# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cosmic Webapp",
    "version": "1.0.0",
    "description": "API documentation with instructions"
  },
  "servers": [
    {
      "url": "/cosmic/"
    },
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/lists": {
      "post": {
        "summary": "Create list",
        "description": "Creates a new list with a name and an optional array of task IDs.",
        "tags": ["Lists"],
        "requestBody": {
          "description": "List creation payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListInput"
              },
              "example": {
                "name": "My List",
                "task_ids": [1, 2, 3]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all lists",
        "description": "Retrieves all lists with their associated tasks (with details).",
        "tags": ["Lists"],
        "responses": {
          "200": {
            "description": "A list of lists with tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ListWithTasks"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/lists/{id}": {
      "get": {
        "summary": "Get list by ID",
        "description": "Retrieves a single list by ID along with its tasks (with details).",
        "tags": ["Lists"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the list",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List with tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "list": {
                      "$ref": "#/components/schemas/ListWithTasks"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "List not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update list",
        "description": "Updates the list name and its associated task IDs.",
        "tags": ["Lists"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the list to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated list data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListInput"
              },
              "example": {
                "name": "Updated List Name",
                "task_ids": [1, 2, 3]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "list": {
                      "$ref": "#/components/schemas/List"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete list",
        "description": "Deletes a list and its associated task links.",
        "tags": ["Lists"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the list to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ACCESS_TOKEN": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "REFRESH_TOKEN": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "List": {
        "type": "object",
        "properties": {
          "list_id": {
            "type": "string",
            "example": "list123"
          },
          "name": {
            "type": "string",
            "example": "My List"
          }
        }
      },
      "ListInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My List"
          },
          "task_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [1, 2]
          }
        },
        "required": ["name"]
      },
      "ListResponse": {
        "type": "object",
        "properties": {
          "list": {
            "$ref": "#/components/schemas/List"
          }
        }
      },
      "ListWithTasks": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/List"
          },
          {
            "type": "object",
            "properties": {
              "tasks": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        ]
      },
      "Task": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "integer",
            "description": "The unique identifier of the task"
          },
          "title": {
            "type": "string",
            "description": "Title of the task"
          },
          "description": {
            "type": "string",
            "description": "Description of the task"
          },
          "lists_id": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "An array of list IDs this task belongs to. A task can belong to multiple lists."
          },
          "status": {
            "type": "string",
            "description": "Current status of the task"
          },
          "dueDate": {
            "type": "date",
            "description": "Due date task"
          },
          "priority": {
            "type": "integer",
            "description": "Priority of task"
          },
          "activity": {
            "type": "object",
            "description": "Related activity of the task",
            "properties": {
              "activity_id": {
                "type": "integer",
                "description": "ID of activity"
              },
              "name": {
                "type": "string",
                "description": "Name of activity"
              }
            }
          },
          "locations": {
            "type": "object",
            "description": "Location information",
            "properties": {
              "location_id": {
                "type": "string",
                "description": "ID of location"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "description": "Creation timestamp"
              },
              "location_name": {
                "type": "string",
                "description": "Name of location"
              },
              "color": {
                "type": "string",
                "description": "Color of location in HEX"
              },
              "description": {
                "type": "string",
                "description": "Description of location"
              },
              "x": {
                "type": "number",
                "description": "X coordinate"
              },
              "y": {
                "type": "number",
                "description": "Y coordinate"
              },
              "z": {
                "type": "number",
                "description": "Z coordinate"
              },
              "keywords": {
                "type": "string",
                "nullable": true,
                "description": "Location keywords"
              },
              "matterport_tag_id": {
                "type": "string",
                "description": "Matterport Tag ID"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the task was created"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the task was last updated"
          }
        }
      },
...
    }
  }
}